feat: update pokemon type
